requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openai-whisper>=20240930
torch==2.6.0
torchaudio==2.6.0
pyannote.audio==3.3.2
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
numpy<2.0.0
